Added LED controls for a GPS Lock and visual blinker.
We wanted a confirmation that GPS lock had been made
before launching, so that was a new PIN and logic.
We also wanted to have a simple blinking LED in time
with the buzzer. We used a separate PIN to keep
the circuitry simpler.

This code was tested on a separate Arduino with Isaac's
test GPS, set for 270.0m, and was triggered by driving
up from 2nd ave to the Blvd of the Allies.

Code could be refactored, but not at the last minute
without further testing possibilities.